Implemented #81
