+++ pyblish_qml/qml/delegates/ErrorDelegate.qml
@@ -33,9 +33,9 @@ BaseDelegate {
                    root.toggle.width -
                    10
 
-            property bool hasLongMessage: item.message.indexOf("\n") != -1 ? true : false
-            property string shortMessage: item.message.split("\n")[0]
-            property string longMessage: item.message
+            property bool hasLongMessage: object.message.indexOf("\n") != -1 ? true : false
+            property string shortMessage: object.message.split("\n")[0]
+            property string longMessage: object.message
 
             Label {
                 text: root.expanded ? body.longMessage : body.shortMessage
@@ -57,19 +57,19 @@ BaseDelegate {
                     model: [
                         {
                             "key": "Filename",
-                            "value": item.fname
+                            "value": object.fname
                         },
                         {
                             "key": "Line Number",
-                            "value": item.line_number
+                            "value": object.line_number
                         },
                         {
                             "key": "Function",
-                            "value": item.func
+                            "value": object.func
                         },
                         {
                             "key": "Exception",
-                            "value": item.exc
+                            "value": object.exc
                         }
                     ]
