+++ pyblish_qml/qml/temp.qml
@@ -1,13 +1,26 @@
-import QtQuick 2.3
-
+import QtQuick 2.0
+import Pyblish 0.1
+import Pyblish.ListItems 0.1 as ListItem
+import "Delegates.js" as Delegate
 
 Rectangle {
-    width: 500
-    height: 500
-
     color: "brown"
+    anchors.fill: parent
 
-    List {
-        model: app.itemModel
+    ListView {
+        anchors.fill: parent
+
+        model: objModel
+
+        delegate: Loader {
+            width: ListView.view.width
+            sourceComponent: ListItem.StandardActions {
+                text: object.name
+                height: 20
+                width: parent.width
+                active: true
+                checked: true
+            }
+        }
     }
 }
